Rework `gen` site generator.
Drop the single yaml file that defines the site in favor of a directory
structure that mirrors the site to be generated.

Created `AbsPath` struct for working with absolute paths.  Created
`ContentFolder` struct to represent a folder in the site content
directory.  Dropped the `Site` struct in favor of `SiteDef`, which holds
all the site definition data and `SiteFile`, which represents the site
metadata file located in the root of the site content directory.

Added initial site data, currently just a partial directory structure
and yaml files with page titles.