New approach to generating HTML using a sequence of tags.
Add a `tags()` method to `Node` and its subclasses that returns a list
of `Tag` objects.  Each `Tag` contains info about a start or end tag or
other node content.

Added a `format_tags()` function to the `html` package that reads the
tags sequence and produces nicely formatted output.

Added a `format` attribute to `Node` and a `Format` enum with `BLOCK`,
`COMPACT` and `INLINE` values.

Added `as_block()` and `as_compact()` helper functions for changing the
default format of a particular element.