Improve error handling for incomplete files.
Enable the unit test for an incomplete `.links` directive at the end of
the file.

Remove the `EOF` terminal from the grammar and drop the `is_eof()`
method and checks from the parser.

Modify the `parse()` method to fail if an error is set during parsing.

Change some production methods so they do not set an error when the
production is optional, for example `paragraph()` and the check for the
`.section` directive in `section_directive()`.