Added `ProductionResult` class for production return values.
Parser productions can have three distinct states: matched, not matched
and errored.  The `ProductionResult` class represents these three
states, and implements `__bool__()`, which returns false for the errored
and not matched states.

Updated the production methods on `Parser` to return a
`ProductionResult` instead of a `bool`.  Since any error is carried
along by the `ProductionResult` object, removed the `error` attribute on
`Parser`.