Update `home()` function.
Change `home()` so that it returns a `Directory` that contains the home
`Page` instead of returning an `IndexPage`.

Update the `Directory` class so that it supports being the root
directory.  Update the `build_document()` method of `Page` so that it
correctly builds the links in the <nav> when faced with a `Directory`
instead of an `IndexPage`.