Updated the `macos_packages/` page.
Updated the header to mention that the installers are now notarized and
include universal binaries.

Added the `nginx` package to the list.

Updated the packages for the latest versions.

Updated the package links for all packages to use GitHub "latest" URLs
in the form `<project>/releases/latest/download/<pkg file name>`.