Add `link()` method to `PageFile` class.
The `link()` method checks the class type of the link and calls `book()`
for a `BoolLink` and `link()` (aliased as `general_link`) for a general
`Link`.

Updated direct calls to `link()` to use `book()` where appropriate.

Modified the signature of the `book()` function to also take an `asin`
parameter and to make the `href` parameter optional.