Added `--check-links` command line option.
Built a simple external link checker.  Added `--check-links` option to
check links instead of generating the site.

Added the `accumulate_links()` method to `Resource` and its subclasses.
The implementation in `Resource` simply throws, and is a no-op in
`File`.  For `Directory`, it calls `accumulate_links()` on all its
children.  For `Page`, it builds its document and finds the `href`s of
all the `<a>` tags.

Added `accumulate_nodes()` method to `Node` to get a list of all nodes
for a `Node` tree.

Added the `hyperlinks` package with helper function
`check_external_links()` that filters the internal links out of the list
of links and makes a test `GET` request to each link.

Added the `requests` package to the project.