Add line number to parser errors.
If the current token is not `None`, use its start line as the line
number of the error.  (This is imperfect: currently we use `None` for
the token value when at EOF, so no line is available when the error is
at the end of the file; similarly the offending token may be on the next
line from the problem in the case of incomplete directives.)

Also added some quoting to highlight directives and modifiers in error
messages.