Replaced the "Bash" page with the "Shell" page.
Removed the `website.pages.bash` package containing the `bash()` page
generation function.

Added a page file for the "Shell" page in `site-src/shell` and added a
home page link to the shell page.  Updated `main_tests.py` to pass with
the changed set of pages.

Fixed a typo in the business novels page.  Reordered some items on the
home page.