Add a `robots.txt` file and generate a sitemap.
Added the `Sitemap` resource class and attach a `Sitemap` to the site
root.  Added helper functions to the `hyperlinks` package for building
the list of URLs for the sitemap: `only_internal_links()`,
`only_page_links()` and `to_absolute_urls()`.

Added an additional method to `Resource`: `build_documents()`.  This is
called after the site resource tree is assembled but before the site is
generated.  This gives the `Page` resource class a chance to build its
final HTML document before generation.