Remove favicons from home page links.
Drop the `favicon` parameter to the `item()` function that builds link
items.  Modify item generation to replace the `<br>` tag with `: ` and
adjust `base.css` to make the subtitle wrap onto a new line as a new
block.

Also remove the `external` parameter from `item()` and drop the
`.external` styles from `base.css`.

Remove the unused favicon files.

In the `generate()` method of `Directory`, pass the `omit_styles` flag
on to children when generating.