Fix bad links and update external links.
Add `find_links()` method to `Page` to accumulate all `<a>` links in the
page.

Add code to `main()` to find links in all pages and print them out.
Create the `URL_UPDATES` dictionary mapping old URLs to new ones and
fold that into the URL map used to update old blog page links to new
ones.

Fix a couple of bad links directly in the XML.