Revamp macOS installer packages page.
Remove favicons from the list.  Split package-specific styles out of
`base.css` into `macos_packages.css`.

Remove the `collection` and `item` classes from the `<ul>` and `<li>`s
for each package, adding the `installer` class only to the package
download link.

Style the `installer` list item to be larger, and give it a "package"
emoji as a list item bullet.