Use `wasm-bindgen` to compile `maze` to web assembly.
Added a makefile in the project root to capture the steps to run
`wasm-pack` on the `maze` project to build for web assembly and copy the
generated `.wasm` and `.js` files to the `site-src` directory.  Added a
temporary `index.html` with embedded Javascript to load the `maze` code,
generate a maze and insert it into a `<pre>` tag.

Updated `get` to include the `maze` directory in the site.