A reasonable first version of the `maze` page.
Added a brief overview and "Implementation" section to the maze page.
Added a title to the "generator" section using the random seed.

Replaced the static HTML index page with the `maze()` function to build
the page in a structured way.

Changed the makefile to rename the `wasm-bindgen` generated `maze.js`
file to `maze_bg.js` and moved the page startup javascript into
`maze.js`.

Moved the CSS for the maze page into `maze.css`.

Added the `Pre` class to the `markup` package as a block element.