Improve terminology for maze grid.
Change the `ElementType` enum to `Element`, since it represents all the
data that belongs to a square, boundary or intersection.

Change the struct previously called `Element` to `Location`, since it
holds both an `Element` (in the new sense) and the x, y coordinates of
the element.

Change the signature of `Grid::set()`, dropping the `x` and `y`
parameters, making `set()` work like `Grid::update()`.  Drop the now
redundant `update()` method.

Update various fields and variables to use the new terminology.

The `Grid` struct now holds a vector of `Element` enums (formerly
`ElementType` enums) instead of `Location`s.