Maze page adds the random seed to its URL.
Update `maze.js` to look for a `?seed=xxx` query parameter and use that
as the random seed if it's valid, otherwise generate a random seed and
update the page URL to add the `?seed=xxx` query param.

Also add a link after the maze to a new maze with a new randomly
generated seed.