Add `omit_styles()` function to `main.py`.
Move the action that omits styles from the generated site out of the
`generate()` method into a "filter" function that's applied to the site
after the documents are built but before they're written to disk.  This
makes the `generate()` method only concerned with file operations.  The
`omit_styles` parameter is removed from `generate()`.

To support the `omit_styles()` function, add the `visit()` method to
`Resource` and its descendants.

Change `main()` so that it exits with a non-zero status if broken links
are found during link checking.

Im `page_file.py`, make the `website.links` imports local to the
`link()` method to break circular imports.