Make `accumulate_links()` a method of `Directory` only.
In `Page`, extract the link finding operation out of
`accumulate_links()` into the `find_links()` method.  Add a base version
of `find_links()` to `Resource` that returns an empty list.

Rework `accumulate_links()` in `Directory` to call `visit()` with an
action that calls `find_links()` on each resource visited.  Remove the
base `accumulate_links()` method in `Resource` and subclasses except for
`Directory`.