Implement stylesheet merging.
Add `merge_stylesheets()` method to `Page` and call it from the
`merge_stylesheets()` function in `main.py`.  Add `__iter__()` and
`insert_after()` methods to `Node` to support walking all the nodes in a
document and replacing `Stylesheet` elements with `Style` elements.  In
`Text`, add a `preformatted` attribute and modify the `markup()` and
`tokens()` methods to skip tokenizing preformatted text.

In the `gemerate()` method of `Resource`, rename the `output_path`
parameter to `output_dir`.

Update `get_options()` to set the (currently) hard-coded `source_dir`
and `output_dir` options, and replace use of these paths with the option
values in `main()`.