Join separate CSS files together when merging.
If a page contains links to multiple CSS files, join their contents
together into one `<style>` block when merging them into the page.

Added `parse_css_file()` function to the `css` package to easily enable
caching of previously parsed files.

Add `__repr__()` and `__add__()` methods to the `CSS` class and change
the `__init__()` method so that a `CSS` instance can be created from a
file path or by giving it a list of parsed rules.

In the `merge_stylesheets()` method of `Page`, accumulate the parsed
stylesheets and join them together before inserting the `<style>` block.