Remove unused CSS rules inside media query blocks.
Add three new methods to the `CSS` class: `reformat()`,
`remove_rules_for_selector()` and `selectors()`.  Also added these
methods on `MediaQuery`.

The `selectors()` method returns the set of CSS selectors found in the
`CSS` object for both top level rules and rules inside media queries.
The `remove_rules_for_selector()` removes any rules that use the given
selector, both at top level and inside media queries.  The `CSS` object
will act on the rules it contains directly and call through to the
corresponding methods on any media queries it contains.

The `reformat()` method on `CSS` calls `reformat()` on each media query,
which will add `WhitespaceToken` nodes to the media query content so
that the serialized output is nicely formatted.  This replaces the
`remove_blank_lines()` method on `MediaQuery`, which worked with the
"unparsed" media query contents.

Added a `parsed` parameter to the `MediaQuery` init method, which will
trigger a call to `parse_rule_list()` when false.  The add method will
check the `parsed` attributes to determine if the resulting media query
contents should be parsed.