Clean up `link_attribute()` and `link_attributes()`.
Change `link_attribute()` to return a pair of strings containing the
attribute name and value.  Change `link_attributes()` to return a list
of attribute name-value pairs.

Modify `book_link()` and `general_link()` to scan the parsed list of
link attribute name-value pairs to extract the list of authors, the date
and the checked flag.