Improve parser structure by pulling error determination up.
Change `general_link_directive()` so that it returns `NotMatched` when
not finding a valid link modifier ("blog", "docs", etc).  Propogate
`NotMatched` up through `general_link()`.  Change `link()` so that it
returns `MissingLinkModifierError` when matching neither a book link nor
a general link, since `link()` begins by matching the `.link` directive,
which requires one of the following productions to be matched.