Add `--find-books` option to `gen`.
The `find_books()` function in `main.py` searches `openlibrary.org` for
matches to all book links in the site and prints out some results when
there are three or fewer matches.

The new `books` package provides the function
`open_library_search_url()` to build the search query and
`search_open_library()` with returns the parsed results as data classes
`OpenLibraryResults` and `OpenLibraryDocs`.