Sketched out the shape of the news extractor.
Added the `extractor.py` main script and `news` package.

The `news` package exposes objects for a news `Item` and an `Items`
collection.  The `Source` class represents a news source, the `Store`
represents a persistent storeage mechanism and the `Cache` represents a
local file cache of items.  The `URL` is a wrapper around a URL string.