Add feed reading to the extractor and display the results.
Add options parsing to the extractor and have it optionally loop
forever, sleeping a specified number of minutes between feed reading.

Update the server to read the cached news items and write them to the
home page.

Update locations in the Docker container, storing the news cache in
/var/news and placing the python files in /usr/local/news.

Update the make file with the extractor files.