Add pagination to the news site.
Created the `Page` class for calculating the number of pages and the
correct slice of news items for a particular page.

Modified the `home()` function to use pages and added the
`numbered_page()` handler to extract the page number from the route.
Added the `all()` function for an unpaginated view.

Inserted "first"/"last" and "next"/"previous" links into the pages.

Changed the `News` class's `__iter__()` method so it returns items in
reverse order.

Fixed a problem in the `build` make target by explicitly specifying the
`source_files` variable as a dependency instead of trying to include it
as part of the `container_files` variable.