Fix custom nginx error pages.
Move custom nginx error pages to `/var/lib/nginx/error_pages` in the
container and to `news/nginx/error_pages` in the project.  Update the
error page styles to match the news pages.

In the Flask server, replace the `404.html` template with the `PAGE_404`
constant.  Make the first/next/last/previous URLs relative.  Rename the
`home.html` template to `news.html`.

Move the static nginx files from `news/nginx/default` to `news/wwwroot`
in the project.  Delete the unused `index.html` and add `robots.txt` and
`sitemap.txt`.

Add a Gunicorn config file to `src` and remove the Gunicorn command line
switches from the `serve` script.  Switch Gunicorn to using a unix
socket, set the log level to `error` and set the `max_requests` option.

Change the location of the Python files in the container from
`/usr/local/news` to `/usr/lib/news` since that's already specified for
the `news` user.

In the `Cache` class, make writing of an updated cache file atomic by
writing to a temp file first and using `rename()` to atomically replace
the existing cache file.

In the `news.html` template (and the custom error pages), make the
margin and padding more consistent on the `<body>`, `<header>`,
`<footer>` and `<main>` elements.

In the nginx config, add locations to match the root and numbered news
pages and error pages.  Add a rewrite rule for numbered pages ending
with a slash.