Add Docker health check.
Add a `HEALTHCHECK` directive to the Dockerfile.  Create an `sbin`
script `check-health` that runs the Python script `health_check.py` to
interrogate container health.

Add a `health` package with a function to check that the expected
processes are running and two stubs for checking that cron jobs have run
and that servers are responding.

Bump the version of `boto3` to the latest patch release.

Add a Makefile target `status` that runs `docker inspect` and shows the
container's status, which includes health data.