Add news health check for extractor job.
Modify `extractor.py` to write a file to the cache directory at the end
of each run.  Add code to the `check_jobs()` function to check that the
modification time on that file is within the last five minutes (the
period of the extractor job).

Add some formatting improvements to the Dockerfile.