In news, fix noisy news.url log message.
When cleaning a URL that contains a fragment but no query, the news
extractor is emitting the warning:

    WARNING:news.url:Value Error parsing "": bad query field: ''

This is being triggered by URLs like

    https://sr.ht/~icefox/oorandom/#a-brief-history-of-random-numbers

Add tests for the `clean_query()` function and check that no log
messages are emitted when the query is an empty string.

Update `clean_query()` to only parse on the non-empty case.