Replace sed with `fillin.py` script in news build.
Using a URL as a replacement value in a sed `s/from/to/g` replacement
expression is problematic since URL characters like `/` and `&` have
special meaning in sed.

For processing the AWS deployment template, replace sed with a new
`fillin.py` script that will replace "{{name}}" variables in the text.