In `News`, use a dict instead of a set for `unique_items`.
In a preparatory step towards updating `Item` objects when they are seen
again, change the `News.unique_items` attribute from a set to a dict,
storing the `Item` object as both key and value.  This is to overcome a
limitation of Python's set type that doesn't allow for the direct
retrieval of a specific item from a set.